[#162] env: SearchInput으로 이름 변경
