+++ src/hooks/atoms.ts
@@ -3,4 +3,7 @@ import { atom } from 'jotai';
 // 댓글 카운트
 const commentCountAtom = atom(0);
 
-export { commentCountAtom };
+// 검색 값
+const searchValueAtom = atom('');
+
+export { commentCountAtom, searchValueAtom };
